feat: Update page titles and metadata across the application to reflect new branding as "NekoHome"
- Changed titles in various pages including Home, About, Contact, FAQ, and Listings to align with new branding.
- Added descriptions for SEO optimization in FAQ and property detail pages.
- Introduced a new Dashboard My Properties page for better property management.
- Implemented a cooldown feature for contact forms to prevent spam submissions.
- Refactored property detail page to separate content into a dedicated component.
- Added TypeScript configuration for improved type safety in the project.
