+++ src/app/(listing)/(grid-view)/property-list/[type]/propertylist.js
@@ -1,15 +1,14 @@
+"use client";
 import DefaultHeader from "@/components/common/DefaultHeader";
 import Footer from "@/components/common/default-footer";
 import MobileMenu from "@/components/common/mobile-menu";
-import PropertyFilteringList from "@/components/listing/list-view/list-v1/PropertyFilteringList";
 
-import React from "react";
+import PropertyFiltering from "@/components/listing/grid-view/grid-full-3-col/PropertyFiltering";
 
-export const metadata = {
-  title: "NekoHome - All List Style",
-};
+import React, { useState } from "react";
 
-const ListV1 = () => {
+const GridFull3Col = ({ params }) => {
+  const [type, setType] = useState(params.type);
   return (
     <>
       {/* Main Header Nav */}
@@ -26,10 +25,18 @@ const ListV1 = () => {
           <div className="row">
             <div className="col-lg-12">
               <div className="breadcumb-style1">
-                <h2 className="title">List View</h2>
+                {params.type == "for-rent" ? (
+                  <h2 className="title">Danh sách bất động sản cho thuê</h2>
+                ) : (
+                  <h2 className="title">Danh sách bất động sản đăng bán</h2>
+                )}
                 <div className="breadcumb-list">
                   <a href="#">Home</a>
-                  <a href="#">For Rent</a>
+                  {params.type == "for-rent" ? (
+                    <a href="#">Cho thuê</a>
+                  ) : (
+                    <a href="#">Đăng bán</a>
+                  )}
                 </div>
                 <a
                   className="filter-btn-left mobile-filter-btn d-block d-lg-none"
@@ -38,7 +45,7 @@ const ListV1 = () => {
                   role="button"
                   aria-controls="listingSidebarFilter"
                 >
-                  <span className="flaticon-settings" /> Filter
+                  <span className="flaticon-settings" /> Lọc
                 </a>
               </div>
             </div>
@@ -48,17 +55,16 @@ const ListV1 = () => {
       {/* End Breadcumb Sections */}
 
       {/* Property Filtering */}
-      <PropertyFilteringList />
-
+      <PropertyFiltering type={type} setType={setType} />
       {/* Property Filtering */}
 
       {/* Start Our Footer */}
-      <section className="footer-style1 pt60 pb-0">
+      {/* <section className="footer-style1 pt60 pb-0">
         <Footer />
-      </section>
+      </section> */}
       {/* End Our Footer */}
     </>
   );
 };
 
-export default ListV1;
+export default GridFull3Col;
