+++ src/app/(property)/(agents)/agents/page.js
@@ -6,7 +6,7 @@ import FilteringAgent from "@/components/property/FilteringAgent";
 import React from "react";
 
 export const metadata = {
-  title: "Agents || Homez - Real Estate NextJS Template",
+  title: "NekoHome - Agents",
 };
 
 const Agents = () => {
@@ -39,8 +39,8 @@ const Agents = () => {
       {/* End Breadcumb Sections */}
 
       {/* Agent Section Area */}
-      <FilteringAgent/>
-      
+      <FilteringAgent />
+
       {/* End Agent Section Area */}
 
       {/* Start Our Footer */}
