+++ src/app/(property)/(dashboard)/add-property/page.js
@@ -7,7 +7,7 @@ import SidebarDashboard from "@/components/property/dashboard/SidebarDashboard";
 import AddPropertyTabContent from "@/components/property/dashboard/dashboard-add-property";
 
 export const metadata = {
-  title: "Dashboard Add Property || Homez - Real Estate NextJS Template",
+  title: "NekoHome - Thêm bài đăng",
 };
 
 const DashboardAddProperty = () => {
