+++ src/app/(property)/(dashboard)/my-properties/page.js
@@ -1,99 +1,11 @@
-"use client";
-import DashboardHeader from "@/components/common/DashboardHeader";
-import MobileMenu from "@/components/common/mobile-menu";
-import Pagination from "@/components/property/Pagination";
-import Footer from "@/components/property/dashboard/Footer";
-import SidebarDashboard from "@/components/property/dashboard/SidebarDashboard";
-import FilterHeader from "../../../../components/property/dashboard/dashboard-my-properties/FilterHeader";
-import PropertyDataTable from "@/components/property/dashboard/dashboard-my-properties/PropertyDataTable";
-import DboardMobileNavigation from "@/components/property/dashboard/DboardMobileNavigation";
-import DefaultHeader from "@/components/common/DefaultHeader";
-import { useState } from "react";
+import DashboardMyProperties from "./myproperty";
 
-
-
-const DashboardMyProperties = () => {
-  const [searchContent, setSearchContent] = useState("");
-  const [min, setMin] = useState(0);
-  const [max, setMax] = useState(5);
-  const [totalPage, setTotalPage] = useState(null);
-  const [sorting, setSorting] = useState("");
-  const page = {
-    min,
-    max,
-    totalPage,
-    searchContent,
-    sorting,
-    setMin,
-    setMax,
-    setTotalPage,
-    setSearchContent,
-    setSorting,
-  }
-  return (
-    <>
-      {/* Main Header Nav */}
-      <DefaultHeader />
-      {/* End Main Header Nav */}
-
-      {/* Mobile Nav  */}
-      <MobileMenu />
-      {/* End Mobile Nav  */}
-
-      {/* dashboard_content_wrapper */}
-      <div className="dashboard_content_wrapper">
-        <div className="dashboard dashboard_wrapper pr30 pr0-xl">
-          <SidebarDashboard />
-          {/* End .dashboard__sidebar */}
-
-          <div className="dashboard__main pl0-md">
-            <div className="dashboard__content bgc-f7">
-              <div className="row pb40">
-                <div className="col-lg-12">
-                  <DboardMobileNavigation />
-                </div>
-                {/* End .col-12 */}
-              </div>
-              {/* End .row */}
-
-              <div className="row align-items-center pb40">
-                <div className="col-xxl-3">
-                  <div className="dashboard_title_area">
-                    <h2>BDS của tôi</h2>
-                    <p className="text">Quản lý danh sách bất động sản</p>
-                  </div>
-                </div>
-                <div className="col-xxl-9">
-                  <FilterHeader page={page} />
-                </div>
-              </div>
-              {/* End .row */}
-
-              <div className="row">
-                <div className="col-xl-12">
-                  <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
-                    <div className="packages_table table-responsive">
-                      <PropertyDataTable page={page} />
-
-                      <div className="mt30">
-                        <Pagination page={page} totalPage={totalPage} />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-              {/* End .row */}
-            </div>
-            {/* End .dashboard__content */}
-
-            <Footer />
-          </div>
-          {/* End .dashboard__main */}
-        </div>
-      </div>
-      {/* dashboard_content_wrapper */}
-    </>
-  );
+// (Optional) SEO & title cho tab trình duyệt
+export const metadata = {
+  title: "NekoHome - Bất động sản của tôi",
+  description: "Trang quản lý các bất động sản đã đăng.",
 };
 
-export default DashboardMyProperties;
+export default function Page() {
+  return <DashboardMyProperties />;
+}
