+++ src/app/layout-wrapper/wrapper.js
@@ -1,3 +1,7 @@
+export const metadata = {
+  title: "NekoHome - Bất động sản",
+  description: "NekoHome - Tìm kiếm và đăng tin bất động sản dễ dàng",
+};
 const Wrapper = ({ children }) => {
   return <>{children}</>;
 };
