+++ src/app/page.js
@@ -2,7 +2,7 @@ import Home from "./home/page";
 import Wrapper from "./layout-wrapper/wrapper";
 
 export const metadata = {
-  title: "Home v1 || Homez - Real Estate NextJS Template",
+  title: "Neko home - Trang chủ",
 };
 
 export default function MainRoot() {
