+++ src/components/property/property-single-style/sidebar/ScheduleTour.js
@@ -1,8 +1,13 @@
 import React, { useState, useEffect } from "react";
 import { apiAuthen } from "@/apis/authen";
 import { apiContact } from "@/apis/contact";
+import { useCooldown } from "@/hooks/useCooldown";
 
 const ScheduleTour = ({ property }) => {
+  const { cooldown, isCoolingDown, startCooldown } = useCooldown(
+    "send-contact",
+    60 * 5000
+  );
   const [role, setRole] = useState(null);
   const [data, setData] = useState(null);
   const [sendContact, setSendContact] = useState({
@@ -58,9 +63,13 @@ const ScheduleTour = ({ property }) => {
       alert("Số điện thoại phải đủ 10 số");
       return;
     }
+    if (isCoolingDown) {
+      return;
+    }
 
     try {
       const res = await apiContact.sendContact(sendContact);
+      startCooldown(true);
       if (res.status === 200) {
         alert("Gửi yêu cầu thành công");
       } else {
@@ -69,13 +78,19 @@ const ScheduleTour = ({ property }) => {
     } catch (error) {
       console.error("Lỗi khi gửi yêu cầu:", error);
 
+      const statusCode = error.response?.status;
       const errorMessage =
         error.response?.data?.message ||
         error.response?.data ||
         error.message ||
         "Đã xảy ra lỗi khi gửi yêu cầu";
-      if (errorMessage === "Contact form submission failed") {
-        alert("Bạn đã gửi yêu cầu về bài đăng này rồi");
+
+      if (statusCode === 402) {
+        alert("Bạn đã gửi liên hệ trước đó");
+      } else if (errorMessage === "Contact form submission failed") {
+        alert("Đã gửi liên hệ trước đó, vui lòng đợi");
+      } else {
+        alert(errorMessage);
       }
     }
   };
@@ -234,7 +249,11 @@ const ScheduleTour = ({ property }) => {
                     className="ud-btn btn-thm"
                     onClick={handleSendContact}
                   >
-                    Gửi yêu cầu
+                    {cooldown
+                      ? `Đã gửi yêu cầu, vui lòng đợi ${Math.floor(
+                          cooldown / 60 / 1000
+                        )} phút ${(cooldown % 60) % 1000} giây nữa`
+                      : "Gửi yêu cầu"}
                     <i className="fal fa-arrow-right-long" />
                   </button>
                 </div>
